another place I forgot to make generic

+++ lib/caterer/action/berkshelf/clean.rb
@@ -15,10 +15,12 @@ module Caterer
           config = env[:config]
           image  = config.images[env[:image]]
 
-          image.provisioners.each do |provisioner|
-            if provisioner.is_a? Caterer::Provisioner::ChefSolo or env[:force_berks_clean]
-              ::Berkshelf.formatter.msg "cleaning Caterer's shelf"
-              FileUtils.remove_dir(shelf, fore: true)
+          if image
+            image.provisioners.each do |provisioner|
+              if provisioner.is_a? Caterer::Provisioner::ChefSolo or env[:force_berks_clean]
+                ::Berkshelf.formatter.msg "cleaning Caterer's shelf"
+                FileUtils.remove_dir(shelf, fore: true)
+              end
             end
           end
